refactor(pro-pay): restructure module
