+++ provider-payment/src/main/kotlin/io/github/wliamp/pro/pay/gtw/IGtw.kt
@@ -1,4 +1,4 @@
-package io.github.wliamp.pro.pay.data
+package io.github.wliamp.pro.pay.gtw
 
 import reactor.core.publisher.Mono
